@@ -1,11 +1,11 @@
-const mongoose = require("mongoose");
-const Expenses = new mongoose.Schema(
-  {
-    name: { type: String, required: true },
-    time: { type: Date, default: Date.now() },
-    price: { type: Number, required: true },
-    userId:{type:mongoose.Schema.Types.ObjectId,ref:"users",required:true}
-  },
-  { timestamps: true }
-);
-module.exports = mongoose.model("expenses", Expenses);
+// import mongoose from"mongoose";
+// const ExpensesModel = new mongoose.Schema(
+//   {
+//     name: { type: String, required: true },
+//     time: { type: Date, default: Date.now() },
+//     price: { type: Number, required: true },
+//     userId:{type:mongoose.Schema.Types.ObjectId,ref:"users",required:true}
+//   },
+//   { timestamps: true }
+// );
+// module.exports = mongoose.model("expenses", Expenses);
